+++ context_chat_backend/network_em.py
@@ -5,13 +5,14 @@
 import logging
 import os
 from collections.abc import Generator
+from time import sleep
 from typing import Literal, TypedDict
 
 import httpx
 from langchain_core.embeddings import Embeddings
 from pydantic import BaseModel
 
-from .types import EmbeddingException, TConfig, TEmbeddingAuthApiKey, TEmbeddingAuthBasic
+from .types import EmbeddingException, RetryableEmbeddingException, TConfig, TEmbeddingAuthApiKey, TEmbeddingAuthBasic
 from .utils import value_of
 
 logger = logging.getLogger('ccb.nextwork_em')
@@ -48,7 +49,7 @@ class ApiKeyAuth(httpx.Auth):
 class NetworkEmbeddings(Embeddings, BaseModel):
 	app_config: TConfig
 
-	def _get_embedding(self, input_: str | list[str]) -> list[float] | list[list[float]]:
+	def _get_embedding(self, input_: str | list[str], try_: int = 3) -> list[float] | list[list[float]]:
 		emconf = self.app_config.embedding
 
 		lengths = [len(text) for text in (input_ if isinstance(input_, list) else [input_])]
@@ -84,13 +85,39 @@ class NetworkEmbeddings(Embeddings, BaseModel):
 					timeout=emconf.request_timeout,
 					auth=auth,
 				)
+				if response.status_code != 200:
+					raise EmbeddingException(response.text)
+		except (
+			EmbeddingException,
+			httpx.RemoteProtocolError,
+			httpx.ReadError,
+			httpx.LocalProtocolError,
+			httpx.PoolTimeout,
+		) as e:
+			if try_ > 0:
+				logger.debug('Retrying embedding request in 5 secs', extra={'try': try_})
+				sleep(5)
+				return self._get_embedding(input_, try_ - 1)
+			raise RetryableEmbeddingException('Error: request to get embeddings failed') from e
+		except httpx.ConnectError as e:
+			if self.app_config.embedding.workers > 0:
+				logger.error(
+					'Error connecting to the embedding server, check if it is running and the logs',
+					exc_info=e,
+				)
+				raise EmbeddingException('Error: failed to connect to the embedding service') from e
+			logger.error('Error connecting to the remote embedding service', exc_info=e)
+			raise EmbeddingException('Error: failed to connect to the remote embedding service') from e
+		except httpx.NetworkError as e:
+			if try_ > 0:
+				logger.debug('Network error while getting embeddings, retrying in 5 secs', extra={'try': try_})
+				sleep(5)
+				return self._get_embedding(input_, try_ - 1)
+			logger.error('Network error while getting embeddings', exc_info=e)
+			raise EmbeddingException('Error: network error while getting embeddings') from e
 		except Exception as e:
-			raise EmbeddingException('Error: request to get embeddings failed') from e
-
-		try:
-			response.raise_for_status()
-		except Exception as e:
-			raise EmbeddingException(f'Error: failed to get embeddings: {response.text}') from e
+			logger.error('Unexpected error while getting embeddings', exc_info=e)
+			raise EmbeddingException('Error: unexpected error while getting embeddings') from e
 
 		# converts TypedDict to a pydantic model
 		resp = CreateEmbeddingResponse(**response.json())