+++ context_chat_backend/utils.py
@@ -100,7 +100,8 @@ def exec_in_proc(group=None, target=None, name=None, args=(), kwargs={}, *, daem
 
 
 def is_valid_source_id(source_id: str) -> bool:
-	return re.match(r'^[a-zA-Z0-9_-]+__[a-zA-Z0-9_-]+: \d+$', source_id) is not None
+	# note the ":" in the item id part
+	return re.match(r'^[a-zA-Z0-9_-]+__[a-zA-Z0-9_-]+: [a-zA-Z0-9:-]+$', source_id) is not None
 
 
 def is_valid_provider_id(provider_id: str) -> bool: