+++ dockerfile_scripts/install_py11.sh
@@ -7,5 +7,5 @@ apt-get update
 apt-get install -y software-properties-common
 add-apt-repository -y ppa:deadsnakes/ppa
 apt-get update
-apt-get install -y --no-install-recommends python3.11 python3.11-venv python3-pip vim git pciutils
+apt-get install -y --no-install-recommends python3.11 python3.11-venv python3-pip
 update-alternatives --install /usr/bin/python3 python3 /usr/bin/python3.11 1