+++ main_em.py
@@ -3,75 +3,101 @@
 # SPDX-FileCopyrightText: 2024 Nextcloud GmbH and Nextcloud contributors
 # SPDX-License-Identifier: AGPL-3.0-or-later
 #
+import logging
 import os
-import signal
-import threading
-import time
+from time import sleep
+from urllib import URL
 
+import httpx
 import uvicorn
-from llama_cpp.server.app import create_app
-from llama_cpp.server.settings import ModelSettings, ServerSettings
-from starlette.datastructures import URL
-from starlette.types import ASGIApp, Receive, Scope, Send
 
-from context_chat_backend.types import TConfig  # isort: skip
 from context_chat_backend.config_parser import get_config  # isort: skip
+from context_chat_backend.logger import get_logging_config, setup_logging  # isort: skip
+from context_chat_backend.ocs_utils import sign_request  # isort: skip
 from context_chat_backend.setup_functions import ensure_config_file, setup_env_vars  # isort: skip
 
 
-last_time_lock = threading.Lock()
-last_time = 0
-holding_cnt = 0
-
-def die_on_time(app_config: TConfig):
-	global last_time
-	while True:
-		time.sleep(60)
-		with last_time_lock:
-			if holding_cnt <= 0 and time.time() - last_time > app_config.embedding.offload_after_mins * 60:
-				print('Killing the embedding server due to inactivity', flush=True)
-				os.kill(os.getpid(), signal.SIGTERM)
+LOGGER_CONFIG_NAME = 'logger_config_em.yaml'
+# todo: config and env var for this
+MODEL_ALIAS = 'em_model'
+STARTUP_CHECK_SEC = 10
+MAX_TRIES = 180  # 30 minutes max
 
 
-class LastAccessMiddleware:
-	'''
-	Records last access time of the request to the embeddings route.
-	'''
-	def __init__(self, app: ASGIApp) -> None:
-		self.app = app
+def _get_main_app_httpx_client_host() -> tuple[httpx.Client, str]:
+	"""Get an HTTPX client to connect to the main app, depending on the deployment type.
 
-	async def __call__(self, scope: Scope, receive: Receive, send: Send) -> None:
-		global last_time, holding_cnt
-		if scope['type'] != 'http':
-			await self.app(scope, receive, send)
-			return
-
-		url = URL(scope=scope)
-
-		if url.path == '/heartbeat':
-			await send({'type': 'http.response.start', 'status': 200})
-			await send({'type': 'http.response.body', 'body': b'OK'})
-			return
-
-		if url.path == '/v1/embeddings':
-			try:
-				with last_time_lock:
-					holding_cnt += 1
-				await self.app(scope, receive, send)
-			finally:
-				with last_time_lock:
-					last_time = time.time()
-					holding_cnt -= 1
+	The second return value is the base URL to use for the main app.
+	Returns
+	-------
+		httpx.Client: The HTTPX client.
+		str: The base URL to use for the main app.
+	"""
+	_headers = {}
+	sign_request(_headers)
+	if os.getenv('HP_SHARED_KEY'):
+		transport = httpx.HTTPTransport(uds=os.getenv("HP_EXAPP_SOCK", "/tmp/exapp.sock"))  # noqa: S108
+		return httpx.Client(transport=transport, headers=_headers), 'main_app'
+	return httpx.Client(headers=_headers), f'http://{os.environ["APP_HOST"]}:{os.environ["APP_PORT"]}'
 
 
 if __name__ == '__main__':
-	# todo
+	# intial buffer
+	sleep(STARTUP_CHECK_SEC)
+
 	setup_env_vars()
 	ensure_config_file()
-
 	app_config = get_config(os.environ['CC_CONFIG_PATH'])
 	em_conf = app_config.embedding
+
+	if em_conf.workers <= 0:
+		print('No embedding workers configured, exiting...', flush=True)
+		exit(0)
+
 	print('Embedder config:\n' + em_conf.model_dump_json(indent=2), flush=True)
+
+	logging_config = get_logging_config(LOGGER_CONFIG_NAME)
+	setup_logging(logging_config)
+	logger = logging.getLogger('emserver')
+	if app_config.debug:
+		logger.setLevel(logging.DEBUG)
+
+	_max_tries = MAX_TRIES
+	_enabled = False
+	_last_err = None
+	client, base_url = _get_main_app_httpx_client_host()
+	# wait for the main process to be ready, check the /enabled endpoint
+	while _max_tries > 0:
+		try:
+			ret = client.get(f'http://{base_url}/enabled')
+			ret.raise_for_status()
+
+			if not ret.json().get('enabled', False):
+				raise RuntimeError('Main app is not enabled, sleeping for a while...')
+		except (httpx.RequestError, RuntimeError) as e:
+			print(
+				f'{MAX_TRIES-_max_tries+1}/{MAX_TRIES}:'
+				f' [Embedding server] Waiting for the main app to be enabled/ready: {e}',
+				flush=True,
+			)
+			_last_err = e
+			sleep(STARTUP_CHECK_SEC)
+			_max_tries -= 1
+			continue
+
+		_enabled = True
+		break
+
+	if not _enabled:
+		logger.error(
+			'Failed waiting for the main app to be enabled. This could indicate an issue with the AppAPI'
+			' Deploy Daemon setup or some issue in the main app setup. Some common causes of the latter'
+			' could be no/no stable internet connection to download the required models, disk space full,'
+			' or this app not being able to contact the Nextcloud server to report progress of the model'
+			' download.',
+			exc_info=_last_err,
+		)
+		exit(1)
 
 	# update model path to be in the persistent storage if it is not already valid
 	if 'model' not in em_conf.llama:
@@ -83,37 +109,42 @@ if __name__ == '__main__':
 			'model_files',
 			em_conf.llama['model'],
 		)
+		logger.debug(f'Trying model path: {em_conf.llama["model"]}')
 
 		# if the model file is still not found, raise an error
 		if not os.path.isfile(em_conf.llama['model']):
 			raise ValueError('Error: Model file not found at the updated path')
 
+	# delayed import for libcuda.so.1 to be available
+	from llama_cpp.server.app import create_app
+	from llama_cpp.server.settings import ModelSettings, ServerSettings
+
 	base_url = URL(em_conf.base_url)
 	host = base_url.components.hostname or base_url.path or '127.0.0.1'
 	port = base_url.components.port or 5000
-
 	server_settings = ServerSettings(host=host, port=port)
-	model_settings = [ModelSettings(model_alias='em_model', embedding=True, **em_conf.llama)]
+	model_settings = [ModelSettings(model_alias=MODEL_ALIAS, embedding=True, **em_conf.llama)]
 	app = create_app(
 		server_settings=server_settings,
 		model_settings=model_settings,
 	)
-	app.add_middleware(LastAccessMiddleware)
 
-	# start the last time thread
-	last_time_thread = threading.Thread(target=die_on_time, args=(app_config,))
-	last_time_thread.start()
-	with last_time_lock:
-		last_time = time.time()
+	uv_log_config = uvicorn.config.LOGGING_CONFIG  # pyright: ignore[reportAttributeAccessIssue]
+	uv_log_config['formatters']['json'] = logging_config['formatters']['json']
+	uv_log_config['handlers']['file_json'] = logging_config['handlers']['file_json']
+
+	uv_log_config['loggers']['uvicorn']['handlers'].append('file_json')
+	uv_log_config['loggers']['uvicorn.access']['handlers'].append('file_json')
 
 	uvicorn.run(
+		# todo: use string import of the app
 		app=app,
 		host=host,
 		port=port,
 		http='h11',
 		interface='asgi3',
-		# todo
-		log_level=('warning', 'trace')[app_config.debug],
+		log_config=uv_log_config,
+		log_level=app_config.uvicorn_log_level,
 		use_colors=bool(app_config.use_colors and os.getenv('CI', 'false') == 'false'),
 		workers=em_conf.workers,
 	)